ui for pocket set up sketched out

+++ src/game/systems.rs
@@ -3,7 +3,7 @@ use bevy_rapier2d::prelude::*;
 
 use crate::{AppState, config::STOPPING_THRESHOLD};
 
-use super::{resources::*, SimulationState, ball::{components::{Ball, CueBall}}, GameSetupState};
+use super::{resources::*, SimulationState, ball::{components::{Ball, CueBall}}, GameSetUpState};
 
 
 pub fn spawn_camera(
@@ -66,32 +66,36 @@ pub fn stopping_threshold(
 
 /// detects if sprites are not moving
 fn balls_not_moving(
-    ball_query: Query<Entity, (With<Ball>, Changed<Transform>)>,
+    // ball_query: Query<Entity, (With<Ball>, Changed<Transform>)>,
+    ball_query: Query<&Velocity, With<Ball>>,
 ) -> bool {
-    ball_query.is_empty() 
+    ball_query.iter().all(|&ball_velocity| {
+        ball_velocity.linvel.length_squared() <= STOPPING_THRESHOLD
+    })
 }
 
 /// Whenever the balls are not moving anymore it handles the transition from the current player to the next.
 /// Moreover, it detects whether the cue ball has been pocketed and sets [`GameSetupState`] to the appropriate state
 pub fn switch_player_condition(
-    ball_query: Query<Entity, (With<Ball>, Changed<Transform>)>, 
+    ball_query: Query<&Velocity, With<Ball>>, 
     cue_ball_query: Query<&Transform, With<CueBall>>,
+    simulation_state: Res<State<SimulationState>>,
     mut active_player: ResMut<ActivePlayer>,
     mut next_app_state: ResMut<NextState<AppState>>,
-    mut next_game_setup_state: ResMut<NextState<GameSetupState>>,
+    mut next_game_setup_state: ResMut<NextState<GameSetUpState>>,
     mut table_status: ResMut<TableStatus>,
 ) {
-    if balls_not_moving(ball_query) {
+    if balls_not_moving(ball_query) && simulation_state.0 == SimulationState::Running {
         active_player.switch_player();
         println!("{:?}", active_player.0);
         next_app_state.set(AppState::GameSetup);
         // If the cue ball is still on the table set the GameSetupState to ShotSetup otherwise set to CueBallSetup
         if cue_ball_query.is_empty() {
-            next_game_setup_state.set(GameSetupState::CueBallSetup);
+            next_game_setup_state.set(GameSetUpState::CueBallSetup);
         } else {
             let cue_ball_transform = cue_ball_query.single();
             table_status.cue_ball_status.initial_position = Some(cue_ball_transform.translation.truncate());
-            next_game_setup_state.set(GameSetupState::ShotSetup);
+            next_game_setup_state.set(GameSetUpState::ShotSetup);
         } 
     }
 }
